+++ changelog.md
@@ -15,13 +15,14 @@
 
 ### Fixes
 * Power on mobs' arrows now scale like the players ones
+* Removed "no pillaring because of monsters" message if engaged mobs is disabled 
 
 ## 4.18.5
 ### Changes
 * Added back the Villager comfort mixin
   * The latest version has a severe crash, so I've reverted this for now
 
-## 4.18.4
+## 4.18.4****
 ### Changes
 * Backported copper armor texture